requirements.txt
fastapi
pydantic
uvicorn
pyyaml
httpx
coloredlogs
pyjwt
